Logged in user sees edit button for only their profile
If you're logged out, the edit link doesn't appear. And
when you're logged in, the edit link only appears next to your
profile.

However! if you're logged in and you know the id of another user,
you can manually put in the /edit/:id and get to edit another
user's profile. That's a problem.