Populate select with usernames
I got rid of the input text for username. We'll just have
users pick from a list of users, because there is no way
I'm going to remember these names. I left the text input
commented out in the template, just in case I want to go back
to it later.